Added Search method for easyCRUD.class.php
Hello, I have been working with Your Db.class.php and this easyCRUD.class.php for quite some time and I have tweaked it a little for my use. No I really want this Useful bit of code to be added in the Class for use for all users of this Extremely useful DB and EasyCRUD class users.
Thank you.